Module Quickjs

module C : sig ... end

type runtime = C.js_runtime Ctypes.structure Ctypes.ptr

JSRuntime represents a Javascript runtime corresponding to an object heap. Several runtimes can exist at the same time but they cannot exchange objects. Inside a given runtime, no multi-threading is supported.

val new_runtime : unit -> runtime

type context = {

rt : runtime;

ctx : C.js_context Ctypes.structure Ctypes.ptr;

}

JSContext represents a Javascript context (or Realm). Each JSContext has its own global objects and system objects. There can be several JSContexts per JSRuntime and they can share objects, similar to frames of the same origin sharing Javascript objects in a web browser.

val new_context : runtime -> context

type value = {

ctx : context;

v : C.js_value Ctypes.structure;

}

val get_exception : context -> string option

val get_exception_exn : context -> string

module Value : sig ... end

val eval : context -> string -> (value, string) Stdlib.result

val eval_once : string -> (value, string) Stdlib.result