Module Number.Prototype

Number.prototype methods

val to_string : ?options:options -> float -> string

to_string ?options n converts number n to string with JS-compatible formatting. Equivalent to JavaScript's Number.prototype.toString().

raises Invalid_argument
if radix is not in range 2-36

val to_fixed : int -> float -> string

to_fixed digits n formats n with digits fractional digits. Equivalent to JavaScript's Number.prototype.toFixed().

raises Invalid_argument
if digits is not in range 0-101

val to_precision : int -> float -> string

to_precision digits n formats n with digits significant digits. Equivalent to JavaScript's Number.prototype.toPrecision().

raises Invalid_argument
if digits is not in range 1-101

val to_exponential : int -> float -> string

to_exponential digits n formats n in exponential notation. Equivalent to JavaScript's Number.prototype.toExponential().

raises Invalid_argument
if digits is not in range 0-101

val to_radix : int -> float -> string

to_radix radix n formats n in the given radix (2-36). Equivalent to JavaScript's Number.prototype.toString(radix).

raises Invalid_argument
if radix is not in range 2-36