Module Quickjs

type runtime

runtime represents a Javascript runtime corresponding to an object heap. Several runtimes can exist at the same time but they cannot exchange objects. Inside a given runtime, no multi-threading is supported.

type context

context represents a Javascript context (or Realm). Each context has its own global objects and system objects. There can be several contexts per runtime and they can share objects, similar to frames of the same origin sharing Javascript objects in a web browser.

type value

value represents a Javascript value which can be a primitive type or an object.

type js_exn = value

type 'a or_js_exn = ('a, js_exn) Stdlib.result

module Value : sig ... end

val new_runtime : unit -> runtime

val new_context : runtime -> context

val eval_once : string -> value or_js_exn

eval_once script

val eval : context -> string -> value or_js_exn

eval context script

val eval_unsafe : context -> string -> value

eval_unsafe context script, you must check exception by yourself

val get_exception : context -> value