Module Quickjs

type runtime

JSRuntime represents a Javascript runtime corresponding to an object heap. Several runtimes can exist at the same time but they cannot exchange objects. Inside a given runtime, no multi-threading is supported.

type context

JSContext represents a Javascript context (or Realm). Each JSContext has its own global objects and system objects. There can be several JSContexts per JSRuntime and they can share objects, similar to frames of the same origin sharing Javascript objects in a web browser.

type value

JSValue represents a Javascript value which can be a primitive type or an object

module Value : sig ... end

val new_runtime : unit -> runtime

val new_context : runtime -> context

val eval : context -> string -> (value, string) Stdlib.result

eval context script

val eval_once : string -> (value, string) Stdlib.result

eval_once script

val get_exception : context -> string option

val get_exception_exn : context -> string